Step 4 and Step 5.
I completed the step 4 and added a loop in main that allows you to choose more options after doing a action so you dont need to run the program again.
I started the step 5 and it worked once, but I don't know why it doesn't work anymore.
I am getting excited about being able to complete the steps, but I also know that my code is messy. I can't wait to finish and refactor it.